% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegDKwm-package.R
\docType{package}
\name{ggsegDKwm-package}
\alias{ggsegDKwm}
\alias{ggsegDKwm-package}
\title{ggsegDKwm: Desikan-Killiany White-Matter Parcellation Atlas for the 'ggsegverse' Ecosystem}
\description{
Axial Desikan-Killiany white-matter parcellation ('wmparc') together with basal-ganglia structures (caudate, putamen, pallidum, thalamus), provided as a 'ggseg_atlas' object for use with 'ggseg'. Cortical grey matter is merged into a single background label. Built in symmetric 'fsaverage_sym' template space with axial slice views.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gbarisano/ggsegDKwm}
  \item Report bugs at \url{https://github.com/gbarisano/ggsegDKwm/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Barisano \email{gbarisano@users.noreply.github.com}

}
\keyword{internal}
